update zabbix
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@37128 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/zabbix/frontends/php/include/classes/user/CWebUser.php
@@ -1,7 +1,7 @@
 <?php
 /*
 ** Zabbix
-** Copyright (C) 2001-2017 Zabbix SIA
+** Copyright (C) 2001-2018 Zabbix SIA
 **
 ** This program is free software; you can redistribute it and/or modify
 ** it under the terms of the GNU General Public License as published by
@@ -29,11 +29,23 @@ class CWebUser {
 	static $set_cookie = true;
 
 	/**
+	 * Flag used to not to extend session lifetime in checkAuthentication.
+	 */
+	static $extend_session = true;
+
+	/**
 	 * Disable automatic cookie setting.
 	 * First checkAuthentication call (performed in initialization phase) will not be sending cookies.
 	 */
 	public static function disableSessionCookie() {
 		self::$set_cookie = false;
+	}
+
+	/**
+	 * Disable automatic session extension.
+	 */
+	public static function disableSessionExtension() {
+		self::$extend_session = false;
 	}
 
 	/**
@@ -76,7 +88,7 @@ class CWebUser {
 			}
 
 			// remove guest session after successful login
-			$result &= DBexecute('DELETE FROM sessions WHERE sessionid='.zbx_dbstr(get_cookie('zbx_sessionid')));
+			$result &= DBexecute('DELETE FROM sessions WHERE sessionid='.zbx_dbstr(get_cookie(ZBX_SESSION_NAME)));
 
 			if ($result) {
 				self::setSessionCookie(self::$data['sessionid']);
@@ -95,15 +107,21 @@ class CWebUser {
 	 */
 	public static function logout() {
 		self::$data['sessionid'] = self::getSessionCookie();
-		self::$data = API::User()->logout([]);
-		CSession::destroy();
-		zbx_unsetcookie('zbx_sessionid');
+
+		if (API::User()->logout([])) {
+			self::$data = null;
+			CSession::destroy();
+			zbx_unsetcookie(ZBX_SESSION_NAME);
+		}
 	}
 
 	public static function checkAuthentication($sessionId) {
 		try {
 			if ($sessionId !== null) {
-				self::$data = API::User()->checkAuthentication(['sessionid' => $sessionId]);
+				self::$data = API::User()->checkAuthentication([
+					'sessionid' => $sessionId,
+					'extend' => self::$extend_session
+				]);
 			}
 
 			if ($sessionId === null || empty(self::$data)) {
@@ -148,16 +166,16 @@ class CWebUser {
 	public static function setSessionCookie($sessionId) {
 		$autoLogin = self::isGuest() ? false : (bool) self::$data['autologin'];
 
-		zbx_setcookie('zbx_sessionid', $sessionId,  $autoLogin ? strtotime('+1 month') : 0);
+		zbx_setcookie(ZBX_SESSION_NAME, $sessionId,  $autoLogin ? strtotime('+1 month') : 0);
 	}
 
 	/**
-	 * Retrieves current session ID from zbx_sessionid cookie.
+	 * Retrieves current session ID from cookie named as defined in ZBX_SESSION_NAME.
 	 *
 	 * @return string
 	 */
 	public static function getSessionCookie() {
-		return get_cookie('zbx_sessionid');
+		return get_cookie(ZBX_SESSION_NAME);
 	}
 
 	public static function setDefault() {
@@ -209,11 +227,34 @@ class CWebUser {
 	}
 
 	/**
+	 * Return true if guest user has access to frontend.
+	 *
+	 * @return bool
+	 */
+	public static function isGuestAllowed() {
+		$guest = DB::select('users', [
+			'output' => ['userid'],
+			'filter' => ['alias' => ZBX_GUEST_USER]
+		]);
+
+		return getUserGuiAccess($guest[0]['userid']) != GROUP_GUI_ACCESS_DISABLED;
+	}
+
+	/**
 	 * Returns refresh rate in seconds.
 	 *
 	 * @return int
 	 */
 	public static function getRefresh() {
 		return timeUnitToSeconds(self::$data['refresh']);
+	}
+
+	/**
+	 * Returns interface language attribute value for HTML lang tag.
+	 *
+	 * @return string
+	 */
+	public static function getLang() {
+		return (self::$data) ? substr(self::$data['lang'], 0, strpos(self::$data['lang'], '_')) : 'en';
 	}
 }
